@@ -1,6 +1,6 @@
 {
   "name": "fast-food-feud-lab",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Folks' Favorite Friendly Fuel Finder For Food Facts",
   "private": true,
   "scripts": {
@@ -9,15 +9,12 @@
     "debug": "vite dev --debug"
   },
   "dependencies": {
-    "instant-noodles": "^0.0.3",
-    "react": "<18.0.0",
-    "react-dom": "<18.0.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@testing-library/react": "12.1.4",
-    "@vitejs/plugin-react": "^1.3.1",
-    "sinon": "^14.0.0",
-    "vite": "^2.9.9"
+    "@vitejs/plugin-react": "^3.1.0",
+    "vite": "^4.1.4"
   },
   "author": {
     "name": "Jeff Astor",
